ClaudeCode-Notion-Monitor/requirements.txt
notion-client==2.2.1
python-dotenv==1.0.0
watchdog==3.0.0
requests==2.31.0
psutil==5.9.6
datetime
asyncio